Add observable-clipboard module
End: Make the workflow more natural.
Means: Remove the Ctrl/Cmd+Shift+C shortcut.

Electron's clipboard API doesn't expose a change event or observable
event stream (see https://github.com/electron/electron/issues/2280).
Until it does, we're forced to busy wait, checking the contents of the
clipboard every N ms.

Fortunately, electron-clipboard-watcher already implements this.
However, it's API isn't ideal. It's as important to start a conversation
with that module's author, @joshwnj, as it is to meet the stated end.

With the above in mind, add the observable-clipboard module, which uses
the wraps the electron-clipboard-watcher with zen-observable, an
implementation of the Observable spec (see
https://tc39.github.io/proposal-observable/).