Conditionally lower file transport's level
By default, electron-log logs >= warn via the file transport and >=
silly to the console transport. Per b1c90794, a packaged application
doesn't have a tty to log to, which makes probing the application's
behaviour difficult.

It'd be useful to have a mechanism to lower the file transport's level
regardless of whether it's part of a packaged application.

Add the log module, which returns the same object returned by

    require('electron-log')

However, if a DEBUG file exists in the root directory, then the file
transport's level is lowered to "silly".